Remove cutscenes causing Black screen issue
There's an issue where triggering the 2 cutscenes in the lower entrance to the white EMMI zone (the CU wake up cutscenes) will cause a black screen (hear audio, but no GUI/camera) if skipped I believe, or maybe when the water is not drained or you skip that part?  I don't know.  Either way, it seems like the OnCutSceneEnd never gets called and that callback contains a simple FadeIn call, and thus, without that... you just get a black screen.
Regardless of the cause or the reasoning, with the Emmi Chase sequence removed, I believe these two cutscenes are now also safe to be removed as they do not trigger any game flags that make any sort of difference.